Register debug-configuration element after the html template is loaded
Previously the element script was pulled in via an HTML import, and
the script registered the element. However, loading the script via the
HTML import bypassed Node's require(), and since the script is a
CommonJS module that did not work... at all.

Now the script is loaded via Node's require(), and the script no longer
registers the element on load because the script is loaded before the
html template, so element registration must be delayed until after the
HTML import of the html template completes.